requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
astunparse=1.6.3=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
blinker=1.8.2=pypi_0
brotli-python=1.0.9=py39hd77b12b_8
ca-certificates=2024.7.4=h56e8100_0
cachetools=5.5.0=pypi_0
certifi=2024.7.4=py39haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
cuda-cccl=12.4.127=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.4.127=0
cuda-runtime=11.8.0=0
cuda-version=11.8=h70ddcb2_3
cudatoolkit=11.8.0=hd77b12b_0
cudnn=8.9.7.29=he6de189_3
deepface=0.0.93=pypi_0
filelock=3.15.4=pypi_0
fire=0.6.0=pypi_0
flask=3.0.3=pypi_0
flask-cors=4.0.1=pypi_0
flatbuffers=24.3.25=pypi_0
freetype=2.12.1=ha860e81_0
gast=0.4.0=pypi_0
gdown=5.2.0=pypi_0
gmpy2=2.1.2=py39h7f96b67_0
google-auth=2.34.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.65.5=pypi_0
gunicorn=23.0.0=pypi_0
h5py=3.11.0=pypi_0
idna=3.7=py39haa95532_0
importlib-metadata=8.3.0=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
itsdangerous=2.2.0=pypi_0
jinja2=3.1.4=py39haa95532_0
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
lcms2=2.16=h67d730c_0
lerc=4.0.0=h63175ca_0
libclang=18.1.1=pypi_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.5.147=0
libcurand-dev=10.3.5.147=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.20=hcfcfb64_0
libjpeg-turbo=3.0.3=h827c3e9_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.6.0=hddb2be6_3
libuv=1.48.0=h827c3e9_0
libwebp-base=1.3.2=h2bbff1b_0
libxcb=1.15=hcd874cb_0
libzlib=1.2.13=h2466b09_6
libzlib-wapi=1.2.13=h2466b09_6
lz4-c=1.9.4=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.7=pypi_0
markupsafe=2.1.5=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.8=py39h2bbff1b_0
mkl_random=1.2.4=py39h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py39haa95532_0
msys2-conda-epoch=20160418=1
mtcnn=0.1.1=pypi_0
networkx=3.2.1=py39haa95532_0
numpy=2.0.1=pypi_0
numpy-base=1.26.4=py39h65a83cf_0
oauthlib=3.2.2=pypi_0
opencv-python=4.10.0.84=pypi_0
openjpeg=2.5.2=hae555c5_0
openssl=3.3.1=h2466b09_2
opt-einsum=3.3.0=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py39haa95532_0
protobuf=3.19.6=pypi_0
pthread-stubs=0.3=h3c9f919_1
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.19=h1aa4202_1
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.9=2_cp39
pytorch=2.4.0=py3.9_cuda11.8_cudnn9_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py39h2bbff1b_0
requests=2.32.3=py39haa95532_0
requests-oauthlib=2.0.0=pypi_0
retina-face=0.0.17=pypi_0
rsa=4.9=pypi_0
setuptools=72.1.0=py39haa95532_0
six=1.16.0=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.12=py39haa95532_0
tbb=2021.8.0=h59b6b97_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.10.1=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.4.0=pypi_0
tk=8.6.14=h0416ee5_0
torchaudio=2.4.0=pypi_0
torchvision=0.19.0=pypi_0
tqdm=4.66.5=pypi_0
typing-extensions=4.12.2=pypi_0
typing_extensions=4.11.0=py39haa95532_0
tzdata=2024.1=pypi_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.2.2=py39haa95532_0
vc=14.40=h2eaa2aa_0
vc14_runtime=14.40.33810=ha82c5b3_20
vs2015_runtime=14.40.33810=h3bf8584_20
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
wrapt=1.16.0=pypi_0
xorg-libxau=1.0.11=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
zipp=3.20.0=pypi_0
zlib=1.2.13=h2466b09_6
zstd=1.5.5=hd43e919_2
